requirements.txt
Flask>=3.0.0
numpy>=1.26.0
pandas>=2.0.0
matplotlib>=3.7.0
scikit-learn>=1.3.0
Pillow>=10.0.0 